Sorting Long arrays instead of Integer arrays

+++ src/sorting/ISortAlgorithm.java
@@ -3,5 +3,5 @@ package sorting;
 import java.util.List;
 
 public interface ISortAlgorithm {
-    public List<Integer> sort(List<Integer> array);
+    public List<Long> sort(List<Long> array);
 }
